# Makes GetChildKeys more efficient