--- a/src/libraries/Microsoft.Extensions.Configuration/src/ConfigurationKeyComparer.cs
+++ b/src/libraries/Microsoft.Extensions.Configuration/src/ConfigurationKeyComparer.cs
@@ -29,39 +29,38 @@ namespace Microsoft.Extensions.Configuration
         /// <returns>Less than 0 if x is less than y, 0 if x is equal to y and greater than 0 if x is greater than y.</returns>
         public int Compare(string? x, string? y)
         {
-            string[] xParts = x?.Split(_keyDelimiterArray, StringSplitOptions.RemoveEmptyEntries) ?? Array.Empty<string>();
-            string[] yParts = y?.Split(_keyDelimiterArray, StringSplitOptions.RemoveEmptyEntries) ?? Array.Empty<string>();
+            if (string.IsNullOrWhiteSpace(x))
+            {
+                if (string.IsNullOrWhiteSpace(y))
+                {
+                    return 0;
+                }
+
+                return -y!.Length;
+            }
+            else if (string.IsNullOrWhiteSpace(y))
+            {
+                return x!.Length;
+            }
+
+            int result;
+            if (!x!.Contains(ConfigurationPath.KeyDelimiter) || !y!.Contains(ConfigurationPath.KeyDelimiter))
+            {
+                result = compare(x, y);
+                if (result != 0)
+                {
+                    // One of them is different
+                    return result;
+                }
+            }
+
+            string[] xParts = x.Split(_keyDelimiterArray, StringSplitOptions.RemoveEmptyEntries);
+            string[] yParts = y.Split(_keyDelimiterArray, StringSplitOptions.RemoveEmptyEntries);
 
             // Compare each part until we get two parts that are not equal
             for (int i = 0; i < Math.Min(xParts.Length, yParts.Length); i++)
             {
-                x = xParts[i];
-                y = yParts[i];
-
-                int value1 = 0;
-                int value2 = 0;
-
-                bool xIsInt = x != null && int.TryParse(x, out value1);
-                bool yIsInt = y != null && int.TryParse(y, out value2);
-
-                int result;
-
-                if (!xIsInt && !yIsInt)
-                {
-                    // Both are strings
-                    result = string.Compare(x, y, StringComparison.OrdinalIgnoreCase);
-                }
-                else if (xIsInt && yIsInt)
-                {
-                    // Both are int
-                    result = value1 - value2;
-                }
-                else
-                {
-                    // Only one of them is int
-                    result = xIsInt ? -1 : 1;
-                }
-
+                result = compare(xParts[i], yParts[i]);
                 if (result != 0)
                 {
                     // One of them is different
@@ -72,6 +71,35 @@ namespace Microsoft.Extensions.Configuration
             // If we get here, the common parts are equal.
             // If they are of the same length, then they are totally identical
             return xParts.Length - yParts.Length;
+
+            static int compare(string? a, string? b)
+            {
+                int value1 = 0;
+                int value2 = 0;
+
+                bool aIsInt = a != null && int.TryParse(a, out value1);
+                bool bIsInt = b != null && int.TryParse(b, out value2);
+
+                int result;
+
+                if (!aIsInt && !bIsInt)
+                {
+                    // Both are strings
+                    result = string.Compare(a, b, StringComparison.OrdinalIgnoreCase);
+                }
+                else if (aIsInt && bIsInt)
+                {
+                    // Both are int
+                    result = value1 - value2;
+                }
+                else
+                {
+                    // Only one of them is int
+                    result = aIsInt ? -1 : 1;
+                }
+
+                return result;
+            }
         }
     }
 }