--- a/src/libraries/Microsoft.Extensions.Configuration/src/ConfigurationKeyComparer.cs
+++ b/src/libraries/Microsoft.Extensions.Configuration/src/ConfigurationKeyComparer.cs
@@ -31,21 +31,50 @@ namespace Microsoft.Extensions.Configuration
         {
             if (x is not null && y is not null)
             {
-                if (x.Contains(ConfigurationPath.KeyDelimiter) && y!.Contains(ConfigurationPath.KeyDelimiter))
+                if (x.Contains(ConfigurationPath.KeyDelimiter) && y.Contains(ConfigurationPath.KeyDelimiter))
                 {
-                    string[] xParts = x!.Split(_keyDelimiterArray, StringSplitOptions.RemoveEmptyEntries);
-                    string[] yParts = y!.Split(_keyDelimiterArray, StringSplitOptions.RemoveEmptyEntries);
+                    int xIndex = 0;
+                    int yIndex = 0;
 
                     // Compare each part until we get two parts that are not equal
-                    for (int i = 0; i < Math.Min(xParts.Length, yParts.Length); i++)
+                    while (xIndex < x.Length && yIndex < y.Length)
                     {
-                        int result = Compare(xParts[i], yParts[i]);
-                        if (result != 0)
+                        int nextXIndex = x.IndexOf(ConfigurationPath.KeyDelimiter, xIndex);
+                        int nextYIndex = y.IndexOf(ConfigurationPath.KeyDelimiter, yIndex);
+
+                        ReadOnlySpan<char> xSpan;
+                        if (nextXIndex >= 0)
                         {
-                            // One of them is different
-                            return result;
+                            xSpan = x.AsSpan().Slice(xIndex, nextXIndex - xIndex);
+                            xIndex = nextXIndex + 1;
+                        }
+                        else
+                        {
+                            xSpan = x.AsSpan().Slice(xIndex, x.Length - xIndex);
+                            xIndex = x.Length;
+                        }
+
+                        ReadOnlySpan<char> ySpan;
+                        if (nextYIndex >= 0)
+                        {
+                            ySpan = y.AsSpan().Slice(yIndex, nextYIndex - yIndex);
+                            yIndex = nextYIndex + 1;
+                        }
+                        else
+                        {
+                            ySpan = y.AsSpan().Slice(yIndex, y.Length - yIndex);
+                            yIndex = y.Length;
+                        }
+
+                        int compareResult = Compare(xSpan, ySpan);
+                        if (compareResult != 0)
+                        {
+                            return compareResult;
                         }
                     }
+
+                    string[] xParts = x.Split(_keyDelimiterArray, StringSplitOptions.RemoveEmptyEntries);
+                    string[] yParts = y.Split(_keyDelimiterArray, StringSplitOptions.RemoveEmptyEntries);
 
                     // If we get here, the common parts are equal.
                     // If they are of the same length, then they are totally identical
@@ -53,23 +82,27 @@ namespace Microsoft.Extensions.Configuration
                 }
                 else
                 {
-                    return Compare(x, y);
+                    return Compare(x.AsSpan(), y.AsSpan());
                 }
             }
 
             return x?.Length ?? 0 - y?.Length ?? 0;
 
-            static int Compare(string? a, string? b)
+            static int Compare(ReadOnlySpan<char> a, ReadOnlySpan<char> b)
             {
+#if NETCOREAPP
                 bool aIsInt = int.TryParse(a, out int value1);
                 bool bIsInt = int.TryParse(b, out int value2);
-
+#else
+                bool aIsInt = int.TryParse(a.ToString(), out int value1);
+                bool bIsInt = int.TryParse(b.ToString(), out int value2);
+#endif
                 int result;
 
                 if (!aIsInt && !bIsInt)
                 {
                     // Both are strings
-                    result = string.Compare(a, b, StringComparison.OrdinalIgnoreCase);
+                    result = a.CompareTo(b, StringComparison.OrdinalIgnoreCase);
                 }
                 else if (aIsInt && bIsInt)
                 {