… src/libraries/Microsoft.Extensions.Configuration/tests/ConfigurationPathComparerTest.cs
… src/libraries/Microsoft.Extensions.Configuration/tests/ConfigurationTest.cs