@davidfowl 2022-03-26
Can you also add a performance test for keys that do need the split?