@maryamariyan 2022-03-26
> Can you also add a performance test for keys that do need the split?

As you suspected the commit was not optimized for the keys needing split. Now with the newest diff I changed it to be optimized for split case. Using:

```c#
        {
// in Setup:
            var splittingKeys = new Dictionary<string, string>() { };
            for (int i = 1000; i < 2000; i++)
            {
                splittingKeys.Add("a:" + i.ToString(), string.Empty);
            }

            _chainedConfigWithSplitting = new ChainedConfigurationProvider(new ChainedConfigurationSource
            {
                Configuration = new ConfigurationBuilder()
                    .Add(new MemoryConfigurationSource { InitialData = splittingKeys })
                    .Build(),
                ShouldDisposeConfiguration = false,
            });
        }

        [Benchmark]
        public void AddChainedConfigurationWithSplitting()
            => _chainedConfigWithSplitting.GetChildKeys(new string[0], null);
```


- Without changes

<img width="957" alt="image" src="https://user-images.githubusercontent.com/5897654/160229623-658bc541-da67-4c10-80c6-ce5ee9149976.png">

- After new commit: optimized for split case

<img width="1111" alt="Screen Shot 2022-03-26 at 1 17 28 AM" src="https://user-images.githubusercontent.com/5897654/160231346-c15763b5-0fc9-44c8-92a7-7499d2587115.png">