@danmoseley 2022-03-26
Should your perf tests get added to the perf repo?

BTW you can copy paste the results tables into an issue like this as they are valid markdown. That’s easier than pasting screenshots.