@maryamariyan 2022-03-28
With new changes,

Without diff:
```
|                                     Method |        Mean |     Error |      StdDev |    Gen 0 |  Gen 1 | Allocated |
|------------------------------------------- |------------:|----------:|------------:|---------:|-------:|----------:|
|         AddChainedConfigurationNoDelimiter |  1,255.4 us |  24.20 us |    50.52 us | 130.8594 | 5.8594 |  1,011 KB |
|               AddChainedConfigurationEmpty | 37,957.7 us | 718.05 us | 1,466.79 us |  76.9231 |      - |  1,010 KB |
|       AddChainedConfigurationWithSplitting |    563.1 us |  11.11 us |    27.88 us |  61.5234 | 5.8594 |    472 KB |
| AddChainedConfigurationWithSplittingLooped |    557.0 us |  11.01 us |    23.69 us |  61.5234 | 5.8594 |    472 KB |

```

With changes:
```
|                                     Method |        Mean |     Error |    StdDev |   Gen 0 |  Gen 1 | Allocated |
|------------------------------------------- |------------:|----------:|----------:|--------:|-------:|----------:|
|         AddChainedConfigurationNoDelimiter |    832.3 us |  16.46 us |  28.39 us | 18.5547 | 2.9297 |    143 KB |
|               AddChainedConfigurationEmpty | 18,357.2 us | 365.24 us | 676.99 us |       - |      - |    143 KB |
|       AddChainedConfigurationWithSplitting |    391.5 us |   7.66 us |  14.01 us |  4.8828 |      - |     40 KB |
| AddChainedConfigurationWithSplittingLooped |    376.5 us |   6.71 us |   9.19 us |  4.8828 | 0.4883 |     40 KB |

```