@danmoseley 2022-03-28
Should your benchmarks above be added to dotnet/performance to protect your change from regressions?