@maryamariyan 2022-03-28
> Should your benchmarks above be added to dotnet/performance to protect your change from regressions?

I plan to add those. Are you recommending to add them before this PR goes in?