@danmoseley 2022-03-28
> I plan to add those. Are you recommending to add them before this PR goes in?

That's your call. Adding them before the main PR (at least say half a day before) has the nice effect of gathering numbers before the change, so you see a nice uptick in the graph when your change goes in, and if it regresses it might possibly be interesting what the original value was. I don't think it matters otherwise.