@maryamariyan 2022-03-29
Found certain use cases for which removing `Split` with `IndexOf` would regress for: e.g. for empty segments, the `Split` API also is taking `RemoveEmptyEntries` option and so a string like `:::` would get and empty parts array.

The replacement function needs to behave similar to `Split`, returning the same count, not place allocation while we traverse one segment at a time to compare against the other string's segment.

One way forward would be to make use of something like the `SpanSplitEnumerator` that was reverted back in this PR https://github.com/dotnet/runtime/pull/37757/files.