@tarekgh 2022-03-29
> Seems like StringTokenizer can do the logic I need, will be using that instead [bd43f55/src/libraries/Microsoft.Extensions.Primitives/src/StringTokenizer.cs](https://github.com/dotnet/runtime/blob/bd43f55230d533af326e727366f61cfd3ed0e050/src/libraries/Microsoft.Extensions.Primitives/src/StringTokenizer.cs)

This looks a good idea to use.