@maryamariyan 2022-03-30
I tried using `StringTokenizer` and that would give us the allocation improvement we wanted but not time perf benefits. So I ended up using an improved version of `IndexOf` for this fix, so we get perf benefits in both time and allocation.


Without diff:

```
|                                     Method |        Mean |     Error |      StdDev |    Gen 0 |  Gen 1 | Allocated |
|------------------------------------------- |------------:|----------:|------------:|---------:|-------:|----------:|
|         AddChainedConfigurationNoDelimiter |  1,255.4 us |  24.20 us |    50.52 us | 130.8594 | 5.8594 |  1,011 KB |
|               AddChainedConfigurationEmpty | 37,957.7 us | 718.05 us | 1,466.79 us |  76.9231 |      - |  1,010 KB |
|       AddChainedConfigurationWithSplitting |    563.1 us |  11.11 us |    27.88 us |  61.5234 | 5.8594 |    472 KB |
| AddChainedConfigurationWithSplittingLooped |    557.0 us |  11.01 us |    23.69 us |  61.5234 | 5.8594 |    472 KB |
```

With latest PR diff:
```
|                                     Method |        Mean |     Error |    StdDev |   Gen 0 |  Gen 1 | Allocated |
|------------------------------------------- |------------:|----------:|----------:|--------:|-------:|----------:|
|         AddChainedConfigurationNoDelimiter |    880.8 us |  17.44 us |  23.88 us | 18.5547 | 2.9297 |    143 KB |
|               AddChainedConfigurationEmpty | 19,589.8 us | 388.99 us | 878.01 us |       - |      - |    143 KB |
|       AddChainedConfigurationWithSplitting |    422.6 us |   7.76 us |  14.95 us |  4.8828 | 0.4883 |     40 KB |
| AddChainedConfigurationWithSplittingLooped |    425.8 us |   8.40 us |  11.49 us |  4.8828 | 0.4883 |     40 KB |
```